fix: surface the vocal pitch playback toggle in the chart preview
The listen-in toggle for charted vocal pitches (issue #10) only existed as
an unlabeled button inside the MIDI editor vocal pitch header, so charters
working in the chart preview could not find it and vocal notes seemed
inaudible during playback. Extract the toggle into a shared component and
show it, labeled, in the vocal track overlay next to the harmony part
selector.

Fixes #58
